feat: local, dev 환경에서는 지원서가  있어도 설문지 삭제 가능

+++ mashup-admin/src/main/java/kr/mashup/branding/facade/application/form/ApplicationFormFacadeServiceImpl.java
@@ -1,19 +1,26 @@
 package kr.mashup.branding.facade.application.form;
 
+import java.util.List;
+
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 
+import kr.mashup.branding.domain.application.Application;
+import kr.mashup.branding.domain.application.ApplicationService;
 import kr.mashup.branding.domain.application.form.ApplicationForm;
 import kr.mashup.branding.domain.application.form.ApplicationFormQueryVo;
 import kr.mashup.branding.domain.application.form.ApplicationFormService;
 import kr.mashup.branding.domain.application.form.CreateApplicationFormVo;
 import kr.mashup.branding.domain.application.form.UpdateApplicationFormVo;
+import kr.mashup.branding.facade.ProfileFacadeService;
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
 public class ApplicationFormFacadeServiceImpl implements ApplicationFormFacadeService {
     private final ApplicationFormService applicationFormService;
+    private final ApplicationService applicationService;
+    private final ProfileFacadeService profileFacadeService;
 
     @Override
     public ApplicationForm create(CreateApplicationFormVo createApplicationFormVo) {
@@ -40,6 +47,17 @@ public class ApplicationFormFacadeServiceImpl implements ApplicationFormFacadeSe
 
     @Override
     public void delete(Long applicationFormId) {
+        // 개발 환경에서는 설문지 삭제 시도시 지원서를 모두 삭제 후 설문지까지 삭제 한다.
+        if (profileFacadeService.isLocal() || profileFacadeService.isDevelop()) {
+            applicationService.deleteByApplicationFormId(applicationFormId);
+            applicationFormService.delete(applicationFormId);
+            return;
+        }
+
+        List<Application> applications = applicationService.getApplicationsByFormId(applicationFormId);
+        if (!applications.isEmpty()) {
+            throw new ApplicationFormDeleteFailedException();
+        }
         applicationFormService.delete(applicationFormId);
     }
 }
